add support for batching and fix documentation
